Cypress/E2E: Move cy.apiCreateBot and fix MM-T251 & MM-T252 (#6471)
* move apiCreateBot and fix MM-T252

* fix MM-T251
