+++ e2e/cypress/integration/mark_as_unread/bot_post_mark_as_unread_spec.js
@@ -11,8 +11,9 @@
 // Group: @mark_as_unread
 
 import {getAdminAccount} from '../../support/env';
+import {beUnread} from '../../support/assertions';
 
-import {markAsUnreadFromPost, beUnread, verifyPostNextToNewMessageSeparator} from './helpers';
+import {markAsUnreadFromPost, verifyPostNextToNewMessageSeparator} from './helpers';
 
 describe('Bot post unread message', () => {
     const sysadmin = getAdminAccount();
@@ -35,18 +36,18 @@ describe('Bot post unread message', () => {
         });
 
         // # Create a bot and get userID
-        cy.apiCreateBot('bot-' + Date.now(), 'Test Bot', 'test bot for E2E test replying to older bot post').then((response) => {
-            const botsUserId = response.body.user_id;
-            cy.externalRequest({user: sysadmin, method: 'put', path: `users/${botsUserId}/roles`, data: {roles: 'system_user system_post_all system_admin'}});
+        cy.apiCreateBot('bot-' + Date.now(), 'Test Bot', 'test bot for E2E test replying to older bot post').then(({bot}) => {
+            const botUserId = bot.user_id;
+            cy.externalRequest({user: sysadmin, method: 'put', path: `users/${botUserId}/roles`, data: {roles: 'system_user system_post_all system_admin'}});
 
             // # Get token from bots id
-            cy.apiAccessToken(botsUserId, 'Create token').then((token) => {
+            cy.apiAccessToken(botUserId, 'Create token').then((token) => {
                 //# Add bot to team
-                cy.apiAddUserToTeam(newChannel.team_id, botsUserId);
+                cy.apiAddUserToTeam(newChannel.team_id, botUserId);
 
                 // # Post message as bot to the new channel
-                cy.postBotMessage({token, channelId: newChannel.id, message: 'this is bot message'}).then((post) => {
-                    botPost = post;
+                cy.postBotMessage({token, channelId: newChannel.id, message: 'this is bot message'}).then((res) => {
+                    botPost = res.data;
                 });
             });
         });
