+++ e2e/cypress/integration/mark_as_unread/mark_as_unread_using_shortcuts_spec.js
@@ -9,7 +9,9 @@
 
 // Group: @mark_as_unread
 
-import {verifyPostNextToNewMessageSeparator, beUnread, switchToChannel, beRead, markAsUnreadFromMenu, showCursor} from './helpers';
+import {beRead, beUnread} from '../../support/assertions';
+
+import {verifyPostNextToNewMessageSeparator, switchToChannel, markAsUnreadFromMenu, showCursor} from './helpers';
 
 describe('Mark as Unread', () => {
     let testUser;
